[FIX] Detections: correction height pixels

+++ Editor/Models/System/systemdetection.h
@@ -58,7 +58,7 @@ public:
     int fieldBot() const;
     void setFieldBot(int f);
     void setCurrentHeightSquares(int v);
-    void setCurrentHeightPixels(int v);
+    void setCurrentHeightPixels(double v);
 
     Map * createDetectionMap();
     void getTargetPosition(QVector3D *position) const;
@@ -84,7 +84,7 @@ protected:
     int m_fieldTop;
     int m_fieldBot;
     int m_currentHeightSquares;
-    int m_currentHeightPixels;
+    double m_currentHeightPixels;
     QHash<Position3D, SystemObject3D *> m_boxes;
 };
